Move all patches to separate file to ensure proper isolation
