+++ src/prime_rl/orchestrator/utils.py
@@ -1,6 +1,5 @@
-from typing import Any, List, Optional
+from typing import Any
 
-import openai.types.chat
 import pandas as pd
 from openai.types.chat.chat_completion import ChatCompletion, Choice
 from openai.types.completion_usage import CompletionUsage
@@ -28,66 +27,6 @@ def get_train_sampling_args(sampling_config: SamplingConfig) -> dict:
     sampling_args["extra_body"]["min_tokens"] = sampling_args.pop("min_tokens")
     sampling_args["extra_body"]["repetition_penalty"] = sampling_args.pop("repetition_penalty")
     return sampling_args
-
-
-def monkey_patch_chat_completion_logprobs():
-    """
-    At large batch sizes and context, constructing OAI's Pydantic model
-    ChatCompletion with logprobs is causing heavy CPU overhead (~200ms per
-    object at 32K context, which translates to >10min overhead at 4K batch
-    size). This function monkey-patches the OAI type and verifiers'
-    post-processing utils to avoid validating the complex logprobs field.
-    """
-
-    class ChoiceAny(Choice):
-        """Same as openai.types.chat.chat_completion.Choice, but without type validation for logprobs field."""
-
-        logprobs: Optional[Any] = None
-
-    class ModdedChatCompletion(ChatCompletion):
-        """Same as openai.types.chat.chat_completion.ChatCompletion, but but using ChoiceAny instead of Choice."""
-
-        choices: List[ChoiceAny]  # type: ignore
-
-    # Patch OAI types
-    openai.types.chat.chat_completion.Choice = ChoiceAny
-    openai.types.chat.chat_completion.ChatCompletion = ModdedChatCompletion
-
-    # Patch verifiers parse_chat_completion_logprobs
-    def patched_parse_chat_completion_logprobs(chat_completion: ModdedChatCompletion) -> list[float]:
-        """Same as verifiers.utils.processing_utils.parse_chat_completion_logprobs, but using arbitrary logprobs type."""
-        assert len(chat_completion.choices) == 1, "Response should always have one choice"
-        assert chat_completion.choices[0].logprobs is not None, (
-            "Logprobs should not be None. Make sure to set logprobs=True in the extra body when making the request to /v1/chat/completions"
-        )
-        assert chat_completion.choices[0].logprobs["content"] is not None, (
-            "Logprob content should not be None. Make sure to set logprobs=True in the extra body when making the request to /v1/chat/completions"
-        )
-        logprobs = [logprob["logprob"] for logprob in chat_completion.choices[0].logprobs["content"]]
-        return logprobs
-
-    # Patch verifiers parse_chat_completion_logprobs
-    def patched_parse_chat_completion_tokens(chat_completion: ModdedChatCompletion) -> list[int]:
-        """Same as verifiers.utils.processing_utils.parse_chat_completion_tokens, but using arbitrary logprobs type."""
-        assert len(chat_completion.choices) == 1, "Response should always have one choice"
-        assert chat_completion.choices[0].logprobs is not None, (
-            "Logprobs should not be None. Make sure to set logprobs=True in the extra body when making the request to /v1/chat/completions"
-        )
-        assert chat_completion.choices[0].logprobs["content"] is not None, (
-            "Logprob content should not be None. Make sure to set logprobs=True in the extra body when making the request to /v1/chat/completions"
-        )
-        tokens = [
-            # tokens are token_id:<int> because we request `return_tokens_as_token_ids` from vllm in GRPOTrainer
-            int(token["token"].split(":")[-1])
-            for token in chat_completion.choices[0].logprobs["content"]
-        ]
-        return tokens
-
-    # Import verifiers here (after patching OpenAI types) so verifiers picks up the patched types
-    import verifiers as vf
-
-    vf.utils.processing_utils.parse_chat_completion_logprobs = patched_parse_chat_completion_logprobs
-    vf.utils.processing_utils.parse_chat_completion_tokens = patched_parse_chat_completion_tokens
 
 
 def parse_num_completion_tokens(responses: list[list[ChatCompletion]]) -> list[int]:
